+++ Plutarch/DataRepr.hs
@@ -15,6 +15,8 @@ module Plutarch.DataRepr (
   I.pdropDataRecord,
   I.DerivePConstantViaData (DerivePConstantViaData),
   I.pasDataSum,
+  I.PConstantData,
+  I.PLiftData,
 
   -- * Fields
   F.PDataFields (ptoFields, type PFields),