+++ Plutarch/DataRepr/Internal.hs
@@ -1,7 +1,7 @@
 {-# LANGUAGE AllowAmbiguousTypes #-}
 {-# LANGUAGE DefaultSignatures #-}
 {-# LANGUAGE UndecidableInstances #-}
-{-# OPTIONS_GHC -Wno-orphans -Wno-redundant-constraints #-}
+{-# OPTIONS_GHC -Wno-orphans #-}
 
 module Plutarch.DataRepr.Internal (
   PDataSum,
@@ -11,7 +11,9 @@ module Plutarch.DataRepr.Internal (
   pdcons,
   pdnil,
   DataReprHandlers (..),
+  PConstantData,
   PDataRecord (..),
+  PLiftData,
   PLabeledType (..),
   type PLabelIndex,
   type PUnLabel,
@@ -25,7 +27,7 @@ module Plutarch.DataRepr.Internal (
 ) where
 
 import Data.Functor.Const (Const (Const))
-import Data.Kind (Type)
+import Data.Kind (Constraint, Type)
 import Data.List (groupBy, maximumBy, sortOn)
 import Data.Proxy (Proxy (Proxy))
 import Data.SOP.NP (cana_NP)
@@ -90,7 +92,17 @@ import Plutarch.DataRepr.Internal.HList (type Drop, type IndexList)
 import Plutarch.Integer (PInteger)
 import Plutarch.Internal (S (SI))
 import Plutarch.Internal.Generic (MkSum (mkSum), PCode, PGeneric, gpfrom, gpto)
-import Plutarch.Lift (PConstant, PConstantRepr, PConstanted, PLift, pconstant, pconstantFromRepr, pconstantToRepr)
+import Plutarch.Lift (
+  PConstant,
+  PConstantDecl,
+  PConstantRepr,
+  PConstanted,
+  PLift,
+  PLifted,
+  pconstant,
+  pconstantFromRepr,
+  pconstantToRepr,
+ )
 import Plutarch.List (PListLike (pnil), pcons, pdrop, phead, ptail, ptryIndex)
 import Plutarch.TermCont (TermCont, hashOpenTerm, runTermCont, tcont, unTermCont)
 import Plutarch.Unsafe (punsafeCoerce)
@@ -479,6 +491,75 @@ instance PIsDataRepr a => PlutusType (PIsDataReprInstances a) where
   pcon' (PIsDataReprInstances x) = pconRepr x
   pmatch' x f = pmatchRepr x (f . PIsDataReprInstances)
 
+{- | Type synonym to simplify deriving of @PConstant@ via @DerivePConstantViaData@.
+
+A type @Foo a@ is considered "ConstantableData" if:
+
+- The wrapped type @a@ has a @PConstant@ instance.
+- The lifted type of @a@ has a @PUnsafeLiftDecl@ instance.
+- There is type equality between @a@ and @PLifted (PConstanted a)@.
+- The newtype has @FromData@ and @ToData@ instances
+
+These constraints are sufficient to derive a @PConstant@ instance for the newtype.
+
+For deriving @PConstant@ for a wrapped type represented in UPLC as @Data@, see
+@DerivePConstantViaData@.
+
+Polymorphic types can be derived as follows:
+
+>data Bar a = Bar a deriving stock (GHC.Generic)
+>
+>PlutusTx.makeLift ''Bar
+>PlutusTx.makeIsDataIndexed ''Bar [('Bar, 0)]
+>
+>data PBar (a :: PType) (s :: S)
+>  = PBar (Term s (PDataRecord '["_0" ':= a]))
+>  deriving stock (GHC.Generic)
+>  deriving anyclass (SOP.Generic, PIsDataRepr)
+>  deriving (PlutusType, PIsData, PDataFields) via PIsDataReprInstances (PBar a)
+>
+>instance
+>  forall a.
+>  PLiftData a =>
+>  PUnsafeLiftDecl (PBar a)
+>  where
+>  type PLifted (PBar a) = Bar (PLifted a)
+>
+>deriving via
+>  ( DerivePConstantViaData
+>      (Bar a)
+>      (PBar (PConstanted a))
+>  )
+>  instance
+>    PConstantData a =>
+>    PConstantDecl (Bar a)
+-}
+type PConstantData :: Type -> Constraint
+type PConstantData h =
+  ( PConstant h
+  , Ledger.FromData (h)
+  , Ledger.ToData (h)
+  )
+
+type PLiftData :: PType -> Constraint
+type PLiftData p =
+  ( PLift p
+  , Ledger.FromData (PLifted p)
+  , Ledger.ToData (PLifted p)
+  )
+
+{- |
+
+For deriving @PConstant@ for a wrapped type represented by a builtin type, see
+@DerivePConstantViaNewtype@.
+-}
+newtype
+  DerivePConstantViaData
+    (h :: Type)
+    (p :: PType) -- The Plutarch synonym to the Haskell type
+  = -- | The Haskell type for which @PConstant is being derived.
+    DerivePConstantViaData h
+
 -- | This uses data equality. 'PEq' instances of elements don't make any difference.
 instance PIsDataRepr a => PEq (PIsDataReprInstances a) where
   x #== y = pdata x #== pdata y
@@ -488,9 +569,14 @@ instance (PIsDataRepr a, All (Compose POrd PDataRecord) (PIsDataReprRepr a)) => 
   x #< y = pto x #< pto y
   x #<= y = pto x #<= pto y
 
-newtype DerivePConstantViaData (h :: Type) (p :: PType) = DerivePConstantViaData h
-
-instance (PIsDataRepr p, PLift p, Ledger.FromData h, Ledger.ToData h) => PConstant (DerivePConstantViaData h p) where
+instance
+  ( PIsDataRepr p
+  , PLift p
+  , Ledger.FromData h
+  , Ledger.ToData h
+  ) =>
+  PConstantDecl (DerivePConstantViaData h p)
+  where
   type PConstantRepr (DerivePConstantViaData h p) = Ledger.Data
   type PConstanted (DerivePConstantViaData h p) = p
   pconstantToRepr (DerivePConstantViaData x) = Ledger.toData x